% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdal-dataset.R
\name{mdal_dataset_group_dataset_count}
\alias{mdal_dataset_group_dataset_count}
\title{Get Dataset Count in Group}
\usage{
mdal_dataset_group_dataset_count(mesh, group = 0L)
}
\arguments{
\item{mesh}{External pointer to MDAL mesh.}

\item{group}{Zero-based index of the dataset group.}
}
\value{
Integer count of datasets.
}
\description{
Returns the number of datasets (timesteps) in a dataset group.
}
