% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdal-dataset.R
\name{mdal_dataset_group_is_scalar}
\alias{mdal_dataset_group_is_scalar}
\title{Check if Dataset Group is Scalar}
\usage{
mdal_dataset_group_is_scalar(mesh, group = 0L)
}
\arguments{
\item{mesh}{External pointer to MDAL mesh.}

\item{group}{Zero-based index of the dataset group.}
}
\value{
Logical: TRUE for scalar data, FALSE for vector (x,y components).
}
\description{
Check if Dataset Group is Scalar
}
