% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdal-dataset.R
\name{mdal_dataset_group_name}
\alias{mdal_dataset_group_name}
\title{Get Dataset Group Name}
\usage{
mdal_dataset_group_name(mesh, group = 0L)
}
\arguments{
\item{mesh}{External pointer to MDAL mesh.}

\item{group}{Zero-based index of the dataset group.}
}
\value{
Character string with group name.
}
\description{
Get Dataset Group Name
}
