% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdal-core.R
\name{mdal_last_status}
\alias{mdal_last_status}
\title{MDAL Last Status}
\usage{
mdal_last_status()
}
\value{
Integer status code (0 = OK)
}
\description{
Returns the status code from the last MDAL operation.
}
