% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdal.R
\name{mdal_mesh_dataset_group_count}
\alias{mdal_mesh_dataset_group_count}
\title{Get Dataset Group Count}
\usage{
mdal_mesh_dataset_group_count(mesh)
}
\arguments{
\item{mesh}{External pointer to MDAL mesh, as returned by \code{\link[=mdal_load]{mdal_load()}}.}
}
\value{
Integer count of dataset groups.
}
\description{
Returns the number of dataset groups in a mesh. Dataset groups contain
temporal or thematic data (e.g., "Depth", "Velocity").
}
\examples{
# PLY file with multiple dataset groups
f <- system.file("extdata/MDAL/tests/data/ply/all_features.ply",
                 package = "rmdal", mustWork = TRUE)
mesh <- mdal_load(f)
mdal_mesh_dataset_group_count(mesh)
}
\seealso{
\code{\link[=mdal_load]{mdal_load()}}
}
