% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdal.R
\name{mdal_mesh_face_count}
\alias{mdal_mesh_face_count}
\title{Get Mesh Face Count}
\usage{
mdal_mesh_face_count(mesh)
}
\arguments{
\item{mesh}{External pointer to MDAL mesh, as returned by \code{\link[=mdal_load]{mdal_load()}}.}
}
\value{
Integer count of faces.
}
\description{
Returns the number of faces in a loaded mesh.
}
\examples{
mfile <- system.file("extdata/dem_with_holes/tnz.adf", package = "rmdal", mustWork = TRUE)
mesh <- mdal_load(mfile)
mdal_mesh_face_count(mesh)
}
\seealso{
\code{\link[=mdal_load]{mdal_load()}}, \code{\link[=mdal_mesh_vertex_count]{mdal_mesh_vertex_count()}}, \code{\link[=mdal_mesh_edge_count]{mdal_mesh_edge_count()}}
}
