% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdal.R
\name{mdal_mesh_projection}
\alias{mdal_mesh_projection}
\title{Get Mesh Projection}
\usage{
mdal_mesh_projection(mesh)
}
\arguments{
\item{mesh}{External pointer to MDAL mesh, as returned by \code{\link[=mdal_load]{mdal_load()}}.}
}
\value{
Character string with projection definition (e.g., WKT, PROJ string),
or empty string \code{""} if undefined.
}
\description{
Returns the coordinate reference system (projection) string for a mesh,
if defined.
}
\examples{
mfile <- system.file("extdata/dem_with_holes/tnz.adf", package = "rmdal", mustWork = TRUE)
mesh <- mdal_load(mfile)
mdal_mesh_projection(mesh)
}
\seealso{
\code{\link[=mdal_load]{mdal_load()}}
}
