% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{mdal_mesh_vertices}
\alias{mdal_mesh_vertices}
\title{Get Mesh Vertices}
\usage{
mdal_mesh_vertices(mesh)
}
\arguments{
\item{mesh}{External pointer to MDAL mesh, as returned by \code{\link[=mdal_load]{mdal_load()}}.}
}
\value{
Numeric matrix with N rows (vertices) and 3 columns (x, y, z).
}
\description{
Returns all vertex coordinates from a mesh as an Nx3 matrix.
}
\examples{
f <- system.file("extdata/MDAL/tests/data/2dm/quad_and_triangle.2dm",
                 package = "rmdal", mustWork = TRUE)
mesh <- mdal_load(f)
v <- mdal_mesh_vertices(mesh)
dim(v)
head(v)
}
\seealso{
\code{\link[=mdal_load]{mdal_load()}}, \code{\link[=mdal_mesh_faces]{mdal_mesh_faces()}}, \code{\link[=mdal_mesh_edges]{mdal_mesh_edges()}}
}
