% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdal-package.R
\docType{package}
\name{rmdal-package}
\alias{rmdal}
\alias{rmdal-package}
\title{rmdal: R Interface to the Mesh Data Abstraction Library (MDAL)}
\description{
Provides bindings to the MDAL (Mesh Data Abstraction Library) for reading and writing unstructured mesh data. MDAL supports numerous formats used in hydrology, meteorology, and numerical modelling including 2DM, UGRID, NetCDF, GRIB, XMDF, Selafin, HEC-RAS, FLO-2D, and many others.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/rmdal}
  \item Report bugs at \url{https://github.com/mdsumner/rmdal/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID})

Other contributors:
\itemize{
  \item Lutra Consulting (MDAL library copyright holder) [copyright holder]
}

}
\keyword{internal}
