% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdal-core.R
\name{mdal_drivers}
\alias{mdal_drivers}
\title{List MDAL Drivers}
\usage{
mdal_drivers()
}
\value{
Data frame with columns: name, long_name, can_read_mesh,
can_write_datasets, can_save_mesh, filters
}
\description{
Returns a data frame of available MDAL drivers and their capabilities.
}
\examples{
mdal_drivers()
}
