% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdal.R
\name{mdal_mesh_driver_name}
\alias{mdal_mesh_driver_name}
\title{Get Mesh Driver Name}
\usage{
mdal_mesh_driver_name(mesh)
}
\arguments{
\item{mesh}{External pointer to MDAL mesh, as returned by \code{\link[=mdal_load]{mdal_load()}}.}
}
\value{
Character string with driver name (e.g., "2DM", "UGRID", "Selafin").
}
\description{
Returns the name of the MDAL driver used to load the mesh.
}
\examples{
f <- system.file("extdata/MDAL/tests/data/2dm/quad_and_triangle.2dm",
                 package = "rmdal", mustWork = TRUE)
mesh <- mdal_load(f)
mdal_mesh_driver_name(mesh)
}
\seealso{
\code{\link[=mdal_load]{mdal_load()}}, \code{\link[=mdal_drivers]{mdal_drivers()}}
}
