% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdal.R
\name{mdal_mesh_extent}
\alias{mdal_mesh_extent}
\title{Get Mesh Extent}
\usage{
mdal_mesh_extent(mesh)
}
\arguments{
\item{mesh}{External pointer to MDAL mesh, as returned by \code{\link[=mdal_load]{mdal_load()}}.}
}
\value{
A named list with elements \code{xmin}, \code{xmax}, \code{ymin}, \code{ymax}.
Values may be \code{NaN} if extent cannot be determined.
}
\description{
Returns the bounding box extent of the mesh in its native projection.
}
\examples{
mfile <- system.file("extdata/dem_with_holes/tnz.adf", package = "rmdal", mustWork = TRUE)
mesh <- mdal_load(mfile)
mdal_mesh_extent(mesh)
}
\seealso{
\code{\link[=mdal_load]{mdal_load()}}, \code{\link[=mdal_mesh_projection]{mdal_mesh_projection()}}
}
