% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdal.R
\name{mdal_mesh_projection}
\alias{mdal_mesh_projection}
\title{Get Mesh Projection}
\usage{
mdal_mesh_projection(mesh)
}
\arguments{
\item{mesh}{External pointer to MDAL mesh, as returned by \code{\link[=mdal_load]{mdal_load()}}.}
}
\value{
Character string with projection definition (e.g., WKT, PROJ string),
or empty string \code{""} if undefined.
}
\description{
Returns the coordinate reference system (projection) string for a mesh,
if defined.
}
\examples{
# Mike21 mesh has CRS defined
f <- system.file("extdata/MDAL/tests/data/mike21/small.mesh",
                 package = "rmdal", mustWork = TRUE)
mesh <- mdal_load(f)
mdal_mesh_projection(mesh)
}
\seealso{
\code{\link[=mdal_load]{mdal_load()}}
}
