% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{mdal_mesh_vertices}
\alias{mdal_mesh_vertices}
\title{Get Mesh Vertices}
\usage{
mdal_mesh_vertices(mesh)
}
\arguments{
\item{mesh}{External pointer to MDAL mesh, as returned by \code{\link[=mdal_load]{mdal_load()}}.}
}
\value{
Numeric matrix with N rows (vertices) and 3 columns (x, y, z).
}
\description{
Returns all vertex coordinates from a mesh as an Nx3 matrix.
}
\examples{
\dontrun{
mesh <- mdal_load("mesh.2dm")
v <- mdal_mesh_vertices(mesh)
head(v)
#>          x        y z
#> [1,] 0.000    0.000 0
#> [2,] 1.000    0.000 0
}
}
\seealso{
\code{\link[=mdal_load]{mdal_load()}}, \code{\link[=mdal_mesh_faces]{mdal_mesh_faces()}}, \code{\link[=mdal_mesh_edges]{mdal_mesh_edges()}}
}
