% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdal-core.R
\name{mdal_version}
\alias{mdal_version}
\title{MDAL Version}
\usage{
mdal_version()
}
\value{
Character string with MDAL version
}
\description{
Returns the version string of the linked MDAL library.
}
\examples{
mdal_version()
}
